Updated views, models, and reportTest.html
In models, got rid of the O option, no one other than student and staff can log in anyways.
In reportTest, commented out the radio selection for categories. This will come from the user that made the report instead.
In views, deleted Other option, it is not needed anymore.
Finally, made the default value of users STU in the database.